# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "QASD - Quality Assurance System Dashboard API",
    "description": "Enterprise Quality Assurance and Document Management System API Documentation",
    "version": "1.0.0",
    "contact": {
      "name": "QASD Development Team",
      "email": "dev@qasd.company.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Development server"
    },
    {
      "url": "https://api.qasd.company.com/api",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ProductionBatch": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Batch ID"
          },
          "number": {
            "type": "string",
            "description": "Batch number"
          },
          "product_id": {
            "type": "integer",
            "description": "Product ID"
          },
          "planned_quantity": {
            "type": "number",
            "description": "Planned quantity"
          },
          "actual_quantity": {
            "type": "number",
            "description": "Actual quantity produced"
          },
          "status": {
            "type": "string",
            "enum": ["planned", "in_progress", "completed", "cancelled"],
            "description": "Batch status"
          },
          "start_date": {
            "type": "string",
            "format": "date-time",
            "description": "Start date"
          },
          "end_date": {
            "type": "string",
            "format": "date-time",
            "description": "End date"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "description": "Error message"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "description": "Validation errors"
          }
        }
      },
      "Success": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "Success message"
          },
          "data": {
            "type": "object",
            "description": "Response data"
          }
        }
      }
    }
  },
  "paths": {
...
    "/production/batches": {
      "get": {
        "tags": ["Production"],
        "summary": "Get production batches",
        "description": "Retrieve paginated list of production batches",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {"type": "integer", "minimum": 1},
            "description": "Page number"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {"type": "integer", "minimum": 1, "maximum": 100},
            "description": "Items per page"
          },
          {
            "name": "search",
            "in": "query",
            "schema": {"type": "string"},
            "description": "Search term"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["planned", "in_progress", "completed", "cancelled"]
            },
            "description": "Filter by status"
          },
          {
            "name": "product_id",
            "in": "query",
            "schema": {"type": "integer", "minimum": 1},
            "description": "Filter by product ID"
          },
          {
            "name": "date_from",
            "in": "query",
            "schema": {"type": "string", "format": "date"},
            "description": "Filter from date"
          },
          {
            "name": "date_to",
            "in": "query",
            "schema": {"type": "string", "format": "date"},
            "description": "Filter to date"
          }
        ],
        "responses": {
          "200": {
            "description": "Production batches retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {"$ref": "#/components/schemas/Success"},
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "batches": {
                              "type": "array",
                              "items": {"$ref": "#/components/schemas/ProductionBatch"}
                            },
                            "pagination": {
                              "type": "object",
                              "properties": {
                                "page": {"type": "integer"},
                                "limit": {"type": "integer"},
                                "total": {"type": "integer"},
                                "totalPages": {"type": "integer"}
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Error"}
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Production"],
        "summary": "Create production batch",
        "description": "Create a new production batch",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["product_id", "planned_quantity", "start_date"],
                "properties": {
                  "product_id": {
                    "type": "integer",
                    "minimum": 1,
                    "description": "Product ID"
                  },
                  "planned_quantity": {
                    "type": "number",
                    "minimum": 0.01,
                    "description": "Planned quantity"
                  },
                  "start_date": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Start date"
                  },
                  "expected_end_date": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Expected end date"
                  },
                  "equipment_id": {
                    "type": "integer",
                    "minimum": 1,
                    "description": "Equipment ID"
                  },
                  "notes": {
                    "type": "string",
                    "description": "Notes"
                  },
                  "materials": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "material_id": {
                          "type": "integer",
                          "minimum": 1
                        },
                        "quantity_required": {
                          "type": "number",
                          "minimum": 0.01
                        }
                      }
                    },
                    "description": "Required materials"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Production batch created",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {"$ref": "#/components/schemas/Success"},
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "batch": {"$ref": "#/components/schemas/ProductionBatch"}
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Error"}
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Error"}
              }
            }
          }
        }
      }
    },
    "/production/batches/{id}": {
      "get": {
        "tags": ["Production"],
        "summary": "Get production batch",
        "description": "Get a specific production batch by ID",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {"type": "integer", "minimum": 1},
            "description": "Batch ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Production batch retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {"$ref": "#/components/schemas/Success"},
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "batch": {"$ref": "#/components/schemas/ProductionBatch"}
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Batch not found",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Error"}
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Error"}
              }
            }
          }
        }
      }
    },
...
  },
...
}